pyproject.toml
[tool.poetry]
name = "db-projects"
version = "0.1.0"
description = ""
authors = ["Ramon Medeiro <r.medeiro10@gmail.com>"]
readme = "README.md"
packages = [{include = "db_projects"}]

[tool.poetry.dependencies]
python = "^3.10"
sqlalchemy = "^2.0.27"
tqdm = "^4.66.2"
psycopg2-binary = "^2.9.9"
python-dotenv = "^1.0.1"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
